+++ cmd/attest.go
@@ -13,7 +13,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/sebastienrousseau/scout/attestation"
+	"github.com/sebastienrousseau/scout-reporting/attestation"
 	"github.com/sebastienrousseau/scout/internal/attest"
 	"github.com/sebastienrousseau/scout/internal/policy"
 	"github.com/sebastienrousseau/scout/internal/report"